# Conflicts:
#	backend/app/src/main/java/ch/zhaw/studyflow/Main.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/AppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/AppointmentManager.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/CalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/CalendarManager.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/MemoryAppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/MemoryCalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/SQLAppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/SQLCalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/curriculum/Degree.java
#	backend/app/src/main/java/ch/zhaw/studyflow/domain/curriculum/DegreeManager.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/AppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/CalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/memory/MemoryAppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/memory/MemoryCalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/sql/SQLAppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/sql/SQLCalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistence/AppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistence/CalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistence/memory/InMemoryAppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistence/memory/InMemoryCalendarDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistence/sql/SQLAppointmentDao.java
#	backend/app/src/main/java/ch/zhaw/studyflow/services/persistence/sql/SQLCalendarDao.java
#	backend/app/src/test/java/ch/zhaw/studyflow/controllers/AuthMockHelpers.java
#	backend/app/src/test/java/ch/zhaw/studyflow/controllers/HttpMockHelpers.java
#	backend/app/src/test/java/ch/zhaw/studyflow/domain/appointment/AppointmentManagerTest.java
#	backend/app/src/test/java/ch/zhaw/studyflow/domain/calendar/CalendarManagerTest.java

+++ backend/app/src/main/java/ch/zhaw/studyflow/controllers/DegreeController.java
@@ -35,7 +35,7 @@ public class DegreeController {
             final HttpResponse response = request.createResponse()
                     .setStatusCode(HttpStatusCode.BAD_REQUEST);
 
-            final Optional<Integer> userId = principal.getClaim(CommonClaims.USER_ID);
+            final Optional<Long> userId = principal.getClaim(CommonClaims.USER_ID);
             if (userId.isPresent()) {
                 request.getRequestBody()
                         .flatMap(body -> body.tryRead(Degree.class))
@@ -60,7 +60,7 @@ public class DegreeController {
             final HttpResponse response = request.createResponse()
                     .setStatusCode(HttpStatusCode.BAD_REQUEST);
 
-            final Optional<Integer> userId = principal.getClaim(CommonClaims.USER_ID);
+            final Optional<Long> userId = principal.getClaim(CommonClaims.USER_ID);
             if (userId.isPresent()) {
                 response.setResponseBody(JsonContent.writableOf(degreeManager.getDegreesForStudent(userId.get())))
                         .setStatusCode(HttpStatusCode.OK);
@@ -78,7 +78,7 @@ public class DegreeController {
             final HttpResponse response = request.createResponse()
                     .setStatusCode(HttpStatusCode.BAD_REQUEST);
 
-            final Optional<Integer> userId = principal.getClaim(CommonClaims.USER_ID);
+            final Optional<Long> userId = principal.getClaim(CommonClaims.USER_ID);
             final Optional<Long> degreeId = requestContext.getUrlCaptures().get("degreeId").map(Long::parseLong);
             if (userId.isPresent() && degreeId.isPresent()) {
                 final Degree requestedDegree = degreeManager.getDegree(degreeId.get());
@@ -101,7 +101,7 @@ public class DegreeController {
             final HttpResponse response = request.createResponse()
                     .setStatusCode(HttpStatusCode.BAD_REQUEST);
 
-            final Optional<Integer> userId = principal.getClaim(CommonClaims.USER_ID);
+            final Optional<Long> userId = principal.getClaim(CommonClaims.USER_ID);
             final Optional<Long> degreeId = requestContext.getUrlCaptures().get("degreeId").map(Long::parseLong);
             if (userId.isPresent() && degreeId.isPresent()) {
                 request.getRequestBody()
@@ -126,7 +126,7 @@ public class DegreeController {
             final HttpResponse response = request.createResponse()
                     .setStatusCode(HttpStatusCode.BAD_REQUEST);
 
-            final Optional<Integer> userId = principal.getClaim(CommonClaims.USER_ID);
+            final Optional<Long> userId = principal.getClaim(CommonClaims.USER_ID);
             final Optional<Long> degreeId = requestContext.getUrlCaptures().get("degreeId").map(Long::parseLong);
             if (userId.isPresent() && degreeId.isPresent()) {
                 final Degree degree = degreeManager.getDegree(degreeId.get());
